move all modeling-assessment functionality into the corresponding folder
