--- src/main/webapp/app/assessment-dashboard/index.ts
@@ -1,2 +1,0 @@
-export * from './assessment-dashboard.component';
-export * from './assessment-dashboard.module';
