+++ src/main/webapp/app/modeling-assessment/conflict.model.ts
@@ -1,4 +1,4 @@
-import { Score } from 'app/entities/modeling-assessment/score.model';
+import { Score } from 'app/modeling-assessment/score.model';
 import { User } from 'app/core';
 import { Feedback } from 'app/entities/feedback';
