+++ src/main/webapp/app/modeling-assessment/index.ts
@@ -1,3 +1,5 @@
+export * from './modeling-assessment-dashboard.component';
+export * from './modeling-assessment-dashboard.module';
 export * from './modeling-assessment.service';
 export * from './modeling-assessment.component';
 export * from './modeling-assessment.route';
