+++ src/main/webapp/app/modeling-assessment/modeling-assessment-dashboard.module.ts
@@ -5,16 +5,17 @@ import { ArTEMiSSharedModule } from '../shared';
 import { JhiLanguageHelper, UserRouteAccessService } from '../core';
 import { HomeComponent } from '../home';
 import { JhiMainComponent } from '../layouts';
-import { AssessmentDashboardComponent } from './assessment-dashboard.component';
+import { ModelingAssessmentDashboardComponent } from './modeling-assessment-dashboard.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { SortByModule } from '../components/pipes/sort-by.module';
 import { ArTEMiSResultModule, ResultComponent, ResultDetailComponent } from '../entities/result';
 import { JhiLanguageService } from 'ng-jhipster';
+import { ModelingAssessmentComponent } from 'app/modeling-assessment/modeling-assessment.component';
 
 const ENTITY_STATES = [
     {
         path: 'course/:courseId/exercise/:exerciseId/assessment',
-        component: AssessmentDashboardComponent,
+        component: ModelingAssessmentDashboardComponent,
         data: {
             authorities: ['ROLE_ADMIN', 'ROLE_INSTRUCTOR', 'ROLE_TA'],
             pageTitle: 'assessmentDashboard.title'
@@ -24,9 +25,9 @@ const ENTITY_STATES = [
 ];
 
 @NgModule({
-    imports: [ArTEMiSSharedModule, RouterModule.forChild(ENTITY_STATES), NgbModule, SortByModule, ArTEMiSResultModule],
-    declarations: [AssessmentDashboardComponent],
-    entryComponents: [HomeComponent, ResultComponent, ResultDetailComponent, AssessmentDashboardComponent, JhiMainComponent],
+    imports: [ArTEMiSSharedModule, RouterModule.forChild(ENTITY_STATES), NgbModule, SortByModule, ArTEMiSResultModule, ModelingAssessmentComponent],
+    declarations: [ModelingAssessmentDashboardComponent],
+    entryComponents: [HomeComponent, ResultComponent, ResultDetailComponent, ModelingAssessmentDashboardComponent, JhiMainComponent],
     providers: [{ provide: JhiLanguageService, useClass: JhiLanguageService }]
 })
 export class ArTEMiSAssessmentDashboardModule {
