+++ src/main/webapp/app/modeling-assessment/modeling-assessment.component.scss
