+++ src/main/webapp/app/modeling-assessment/modeling-assessment.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
 import { SERVER_API_URL } from 'app/app.constants';
-import { Result } from '../result';
+import { Result } from '../entities/result';
 import { UMLModel, ElementType, UMLElementType, UMLRelationshipType, UMLClassifier } from '@ls1intum/apollon';
 import { Feedback } from 'app/entities/feedback';
