+++ src/main/webapp/app/modeling-assessment/score.model.ts
